Read a list row's own controls, not the whole page's
The spec helpers used to search the whole page for a row's arrow, its way
in, and a form's button words, so a control rendered against the wrong row
— or words whose button was taken away — still passed.

- form-controls.ts is split into reading (what a page offers), rules
  (whether a value could really be sent), and a thin sending file, with
  every importer pointed at the file that owns what it uses.
- A list is now read into rows, each known by the link that names it, and
  the arrows, markers, and ways in are read off that row's markup alone.
  Statuses, site pages, and API keys all open their lists this way.
- A form found by its body text alone only counts when it has no buttons
  at all — words on a form whose buttons all say something else are no
  button anybody could press, and the send is refused.

Closes the last four TODO sections, which were one defect. Also re-points
a stale equivalent-mutant entry for the button-value fallback.
