+++ TODO.md
@@ -1777,104 +1777,3 @@ Two directions, either of which would help:
   assuming it is.
 
 The first is the real fix; the second would catch the next one either way.
-
-## Split the form-control rules into files about one thing each
-
-*Origin: Codex review on PR #2025. Attempted on that branch and backed out —
-see below.*
-
-`test/specs/support/form-controls.ts` is 478 lines, over the ~400 the repo
-asks for, and holds four separate jobs:
-
-- reading a page's attributes (`attribute`, `hasFlag`, `usableInputsOfKind`)
-- what a page offers (`chooserFor`, `boxFor`, `choicesOffered`, the checkbox
-  and question readers)
-- why a value could not be sent (`whyValueCannotBeSent` and the rules under
-  it, plus the insisted-control machinery)
-- the story-facing helpers (`fillInAndSend`, `takeDownFromActions`)
-
-The first three are pure and the last does the sending, so the natural shape
-is `form-controls/reading.ts`, `form-controls/rules.ts`, and a thin
-`form-controls.ts` — the same split already done for `test-browser.ts`.
-
-The churn is smaller than it looks: `fillInAndSend` has 15 importers and stays
-put, and every reader that would move has between one and five
-(`checkboxValueOffered` 5, `tickedCheckboxes` 4, `whyValueCannotBeSent` 3,
-`requireCheckboxOffered` 2, `choicesOffered`/`optionsOffered` 1 each). So
-about a dozen import lines change. Do not add a re-export layer in
-`form-controls.ts` to avoid touching them — that is the alias-export smell the
-repo rules out; point each caller at the file that owns what it uses.
-
-**Why it was backed out:** attempted by slicing the file on line ranges, which
-produced an unterminated comment, duplicated imports and several unresolved
-symbols. Reverted rather than pushed half-done. Whoever picks this up should
-move whole declarations (or use an editor that understands the syntax) rather
-than cutting on line numbers, and lean on `deno task precommit` — the 214
-specs and the coverage gate both exercise this module hard.
-
-## A form found by its words alone can be sent with no button to press
-
-*Origin: Codex review on PR #2025. Real, and deliberately left for its own
-change — see the sweep below.*
-
-`findFormByButton` picks a form when the button's words appear anywhere in its
-body, then asks `buttonToPress` for the button. When no button matches but the
-words do, `buttonToPress` returns `{}` — "no button with that text at all" —
-and the form is submitted anyway, with no button data.
-
-That is on purpose for forms found by their body text, and plenty are. But it
-means a form whose button is *removed* still submits if the words survive
-elsewhere in it. Site-page deletion is exactly that shape: the heading and the
-button both say "Delete Page", so deleting the button leaves the heading, and
-the story goes on deleting pages the owner has no control to delete.
-
-The fix is not one line. Refusing every no-button case would break every story
-that legitimately finds its form by body text, so the change is to tell those
-two situations apart — probably by having the caller say which it expects, or
-by only allowing the body-text match when the form has no buttons at all.
-Either way it needs a sweep of all 215 scenarios to see which rely on which.
-
-## An arrow is found across the whole page, not on its own row
-
-*Origin: Codex review on PR #2025, raised twice. The second raise carried a
-case the first did not, which is why it is here rather than declined.*
-
-`canMove`/`move` in `test/specs/support/reordering.ts` look for a row's
-`/id/move-up` address anywhere on the page. If that form is rendered against
-the wrong row — present, but beside a different item — the story still submits
-it and passes, while the organiser looking at the named row sees no arrow.
-
-My first answer to this was that a positive scenario would catch an address
-convention changing, which is true but only covers one defect. A form
-*relocated* to another row keeps every address the template tests assert, so
-nothing catches it.
-
-Closing it means attributing controls to rows: parse the list into rows and ask
-what each row offers, rather than searching the page. `openAtState` in
-`statuses.ts` already holds the matched row, so the shape exists — the work is
-giving the shared reordering helper the same scope, for every list that uses it
-(states, site pages).
-
-### The way *into* a row is found the same way
-
-*Origin: a third Codex raise, on PR #2025, against `findsTheWayInFrom`.*
-
-`findsTheWayInFrom` in `test/specs/support/browser.ts` searches
-`row.browser.links` — every link on the page, not the matched row's. Its three
-callers all match on something a sibling row could carry:
-
-| Caller | What it matches | Its `openAt` gives |
-| --- | --- | --- |
-| `statuses.ts` | `href === /statuses/{id}` | the row's markup |
-| `site-pages.ts` | `href` matching `/pages/{id}` | the row's markup |
-| `api-keys.ts` | link text plus an address pattern | no row at all |
-
-Same defect as the arrow above: a link rendered against the wrong row still
-satisfies the search, so a deletion journey passes while the person looking at
-that row has no way in.
-
-Do it with the arrow, not before it. Two of the three callers already hold the
-matched row, but `api-keys.ts` has no row concept yet, so a real fix has to give
-every list the row-parsing shape — which is the same mechanism the arrow needs.
-Fixing the two that are easy would leave a helper whose scoping depends on which
-caller you came from, which is worse than the page-wide search it replaced.
