+++ test/specs/support/editors.ts
@@ -24,10 +24,8 @@ import {
   opensPagesAs,
   rememberBrowser,
 } from "#test/specs/support/browser.ts";
-import {
-  expectCanReallySend,
-  fillInAndSend,
-} from "#test/specs/support/form-controls.ts";
+import { expectCanReallySend } from "#test/specs/support/form-controls/rules.ts";
+import { fillInAndSend } from "#test/specs/support/form-controls.ts";
 import {
   listingIdNamed,
   listingNamed,
