+++ test/specs/support/listings.ts
@@ -12,10 +12,8 @@ import { expect } from "@std/expect";
 import { requireListingWithCount } from "#shared/db/listings/records.ts";
 import type { Listing } from "#shared/types.ts";
 import { adminBrowser, openAdminPage } from "#test/specs/support/browser.ts";
-import {
-  fillInAndSend,
-  requireCheckboxOffered,
-} from "#test/specs/support/form-controls.ts";
+import { requireCheckboxOffered } from "#test/specs/support/form-controls/reading.ts";
+import { fillInAndSend } from "#test/specs/support/form-controls.ts";
 import { minorUnits } from "#test/specs/support/money.ts";
 import type { TicketsWorld } from "#test/specs/support/world.ts";
 import { createTestListing } from "#test-utils/db-helpers/listings.ts";
